Fix for JRUBY-3059: load "foo.rb" should search for .class as well
In order to make it easier to precompile .rb scripts and still have 'load' work without ugly either/or rb/class tricks, I've come around to the idea that load should try both when you request to load a .rb file. This also addresses one key problem with Rails, where its load process is much more conscripted, using its own load path and file-searching magic in concert with "load" instead of "require". Because a patch was needed to make this process accept precompiled .rb files, and because precompiled should generally be interchangeable with uncompiled, I've modified load to search both when a .rb file is specified.

git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@7919 961051c9-f516-0410-bf72-c9f7e237a7b7