A crummy fix for JRUBY-3057: External command spawning in multiple concurrent threads can go boom
This is a fix for the "actual" problem of many threads trying to require the path_helper library at the same time. In short, it synchronizes on the LoadService to prevent multiple threads from trying the same require at the same time. Barring 1.1.5 including a more robust concurrent require fix, this is arguably safer behavior and certainly fixes the problem.

git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@7920 961051c9-f516-0410-bf72-c9f7e237a7b7