Fix for JRUBY-3155: TCPSocket#puts block when the socket is waiting in read
The fix here was to make ChannelStream#feof not be synchronized. It only checks a field on the stream now, rather than trying to do a test read of a single character. It also does not modify state, so it does not need to be synchronized. The problem arose because most of the write operations check for a read EOF first, which meant that if read was blocking, it held the monitor for the stream, and feof would also block. Removing synchronization is safe in this case, and allows the provided script to run to completion.

git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@8082 961051c9-f516-0410-bf72-c9f7e237a7b7