Fix for JRUBY-3195: jruby-openssl fails to load but reports no error
I modified the load logic to no longer swallow all errors, or even LoadError; it only swallows Gem::LoadError when trying to *activate* the gem. It allows errors during jruby-openssl requiring to propagate, as well as errors in setting up that gem or the dummy wrapper. This should correct all cases where errors cause jruby-openssl to fail loading and not tell you why.

git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@8209 961051c9-f516-0410-bf72-c9f7e237a7b7