Fix for JRUBY-2948: Exceptions do not cut off at binding evals after change reported in JRUBY-2945
This problem arose because I had nowhere to stick the "binding frame" flag on StackTraceElement, and so had to disable the trace-shortening logic that cuts off backtraces at a binding/eval frame. My fix was to create a new type RubyStackTraceElement that wraps StackTraceElement and includes a field for "binding frame", allowing me to reenable the cutoff logic. Traces in e.g irb look fine now (clean, short), and I don't think I added back a whole lot of overhead.

Stacktraces are still a nasty bit of code that needs to be revisited and replaced some day. Some day.

git-svn-id: http://svn.codehaus.org/jruby/trunk/jruby@8234 961051c9-f516-0410-bf72-c9f7e237a7b7