Fix for JRUBY-3217: Memory Leak with Adopted Threads
Though we had weakrefs in many places relating to adopted threads, we kept one hard reference in NativeThread. This resulted in thread objects not going away, and since they did not go away, all the JRuby runtime data associated with them did not go away either. The fix makes the reference in NativeThread a weak reference, and deals with it possibly being collected while NativeThread is still alive and referenced (for example, if the actual ruby Thread object is still referenced).

git-svn-id: http://svn.codehaus.org/jruby/tags/jruby-1_1_6RC2@8274 961051c9-f516-0410-bf72-c9f7e237a7b7