Fix for JRUBY-3647: Severe performance degradation from 3d9140fafcda9c4fe6b9d5a1fec0ae9822877e03
My lock-free approach to fixing JRUBY-3551 was clever, but introduced too much overhead into applications that do a lot of includes at runtime (usually in the form of obj.extend). So this falls back on the second-based fix, introducing a global hierarchy lock for when multiple threads concurrently tweak the hierarchy bits. Should still be as safe as before, only slightly slower for hierarchy modifications, and way faster for creating new classes or including modules.