Fix for JRUBY-3714: ActiveMessaging poller stops silently with JRuby 1.3.0RC1 and RC2.
The problem here was that when we built a brand new static scope to contain a Thread's backref/lastline values, we were not entirely copying the original scope. Specifically, we were not using the original scope's previousCRefScope, instead allowing the scope to re-determine what cref scope to use. It ended up choosing the wrong one the second time through, resulting in cvars being incorrectly looked up. Modifying Thread proc creation to also copy the original cref scope fixed the issue.