Rework of Generator library to be 100% ruby again.
This is an attempt to move Generator back to Ruby where it will be easier
to implement lightweight enumerators/generators that do not require threads.
The previous implementation was split between code in generator.rb and the
Generator library in Generator.java. This new version is entirely contained
in generator.rb, offering the possibility of having custom generators for
core classes and an easier path to implementing the lightweight generator
protocol.