Multiple improvements for JRUBY-4117: RubyIO#incrementLineno is Very Slow
I modified both $< and $. to be normal fields on org.jruby.Ruby, with the global var accessors just accessing them directly. This allows all places in the system that access them to go straight in rather than doing the global var hash lookup and constructing a RubyFixnum. I also modified the RubyArgsFile.setCurrentLineNumber to not dispatch back to lineno= and instead just update the int in the ArgsFileData structure. With these changes, we are trimming off the couple seconds RJ Lorimer mentioned in JRUBY-2810.