Fix for JRUBY-4271: interfaces on reified classes patch was incomplete and broke instances being passed to java code.
The problem here was that interface implementing become_java!-fied classes still used the default toJava for Java integration, which ultimately returned just RubyObject or else errored out. Adding a toJava override that returns the actual Java class fixes the problem.

No tests or specs for this, but there's none for become_java in general yet. Need to remedy that.