Unzip native libs duirng the build.
So that they could be properly used when JRuby
is being invoked right from the repository.

Before this commit, we were unpacking the native libs
only during the 'dist' target. So, if JRuby was used right
from the repo directly it didn't see the libs and extracted them
every time from the jars during the startup.

Use 'ant clean-all' to remove those extracted libs, if needed.