Fix for JRUBY-4504: Installing gems while in 1.9 mode fails
The problem here was that we fixed read_nonblock to raise EAGAIN on empty string (meaning no data available), which is correct, but under 1.9 caused net/http to blow up. In 1.9, net/protocol does read_nonblock to fetch URLs, which does often cause EAGAIN to be raised. But we were missing the 1.9 logic that *extends* the EAGAIN error with IO::WaitReadable before raising it, and net/protocol was rescuing WaitReadable. So the EAGAIN went all the way out and caused peculiar results as RubyGems failed over to old indexes and eventually gave up.