New object_id implementation with uniqueness guarantees.
Because identityHashCode turned out to be very unreliable as a unique ID, I redesigned object_id logic to use a strictly-increasing AtomicLong for both ObjectSpace and non-ObjectSpace modes. The long is stored into our object variable table, which is lazily allocated. This does mean that when object_id is called on a given object, the parent's variable table size increases. However, this size increase only affects other objects of the same type that also have object_id called on them (as you would expect) or which have other instance variables (in which case the size increase is just sizeof(pointer) for the extra slot). In general, this should not increase memory usage substantially except for objects that have had object_id called. No additional memory is used for objects that never have object_id called against them or same-typed objects. Also, because the new object_id is based on a Long * 2, object_id is guaranteed to be unique up to 2^63.