Remove old unused Ruby-Init feature.
This feature added logic to the requiring of jar files that looked for a Ruby-Init entry in the jar's manifest. If found, it would run that script as well as adding the jar to the classloader.

The feature does not appear to have ever been used. A Google code search turned up only hits in our own repository. Given that there's separate logic for finding a Service within the jar (ExtensionSearch, I think this is safe to delete. We can reinstate it or another feature once the load process has been cleaned up and simplified.