Sweater unraveling for JRUBY-4615: java integration specs interference
The original problem here was that depending on what order JI-related code was used, a different object could get cached for a java.lang.Class in ObjectProxyCache. Normally, it should just be the proxy wrapper, but some paths inserted a the RubyClass version of the class. This led to peculiar failures if specs were reordered.

The fixes made here are as follows:

* Eliminate all remaining to_java_object logic, which is where the bad OPC.put originated from
* Move all coerce*ToJavaObject logic from JavaUtil into the appropriate toJava overrides
* Fixes and modifications surrounding the above two changes