Modify rdoc to load from a jar file and a marshalled cache, installed in "share/" during the first build.
This new logic allows us to ship without the thousands of files from system-level rdoc dumped directly onto the users filesystem. Instead, they're inside a jar and our File logic for jar files gets used to access them. To blunt the cost of searching the jar file for docs, a cache is generated and saved. The combination means that rdocs take as much as 10x less space on disk and actually appear to load a bit faster. There are other intangible benefits like reduced time to search a JRuby codebase, since the rdocs aren't splattered in there too.