Fix for JRUBY-4764: Leaking fileno to ChannelDescriptor mappings in Ruby.retainedDescriptors.
This change remains on master, but I'm reverting it on 1.5 because the old weak map is a known quantity that worked for years. We'll continue to explore a hard map + finalizers on master for 1.6, and I'll file a bug for that work.

Revert "No need to use a weak map for fileno to ChannelDescriptor, since in C land unclosed fd's live forever too. Eliminates special need for a hard map for sysopen and weak reference management for all other IOs."

This reverts commit 0c012c78053e9ec19ffcdd48231a1141aba544b5.

Conflicts:

	src/org/jruby/Ruby.java