Adapt Socket.getaddrinfo and Socket.getnameinfo for IPv6 addresses.
Note that InetAddress.getAllByName(null) returns Inet4Address[] without hints, so, when necessary, we use "[::]" and "[::1]" to ensure that we get Inet6Address[].

In theory, this should fix a couple of Socket.getaddrinfo specs for IPv6 addresses, but it does not, since the specs specifically asks for "::" and "::1", but InetAddress.getAllByName returns "0:0:0:0:0:0:0:0" and "0:0:0:0:0:0:0:1", respectively. This should be fixed in RubySpec.