Fix for JRUBY-4767: JRuby and open-uri File handle issue
The fix here is to no longer hold a hard reference to a Selector in each RubyThread, since on some environments (like Tomcat) RubyThread objects can build up a bit. Even a modest number of RubyThread objects will eat 3*N file descriptors for epoll selection, which quickly hits the ulimits on a typical Linux box. By opening and closing them for each select, we pay a small per-select hit, but it fixes the problem.