Fix JRUBY-4982: jruby embeds outdated timezone data via joda-time
Adds the ant task 'update-tzdata' and the tasks it depends on. 'update-tzdata' is a completely independent task, so it does not disrupt the current release process. So a call to the 'jar' task is still required for JRuby to see the new tzdata.

As described in http://ant.apache.org/manual/Tasks/ftp.html you'll need jakarta-commons-net-1.4.0 or later in a place ant can find.