Fix for JRUBY-5056: Time marshal format differs between Ruby and JRuby
The problem here was that MRI normalizes time to UTC for the dump format, and the "gmt" bit is only to indicate whether it should be left in that format or localized on load. Modified our dump and load to assume UTC on the wire and switch to local after loading if indicated by the "gmt" bit.