Fix for JRUBY-5057: error from ManyVarsDynamicScope.java after calling jruby->java->jruby
The problem was that EmbedRubyRuntimeAdapterImpl.getManyVarsDynamicScope was pushing a scope that never got popped. When running in a mixed-up environment (like ruby => java => ruby) this could cause a scope to get pushed onto the parent runtime's stack that never got popped. As a result, subsequent scoping on that thread was broken. Fixed by not pushing any scope, and just using a dummy scope locally.