Fix JRUBY-5221: jruby.reify.classes raising error with optparse
The problem here was that the class in question (OptionParser::Switch) had a no-arg "pattern" method on both the object and the class. When generating reified Java methods, this caused an error, since it is not allowed to have both a static and an instance method with the same name and signature. I fixed it by giving priority to the instance method, since normally those will be the most important for reification (for interface impl, etc).

* Prefer instance methods when name+signature conflicts with a static method
* Move jitted code to rubyjit top-level package and reified classes to rubyobj top-level package
* Add some debug logging to reify logic