Fix JRUBY-5225: New method-name-mangling may not be JVM specification compliant
It turns out that there is an amendmend to the classfile format in 1.5 that's not in the normal specification: http://java.sun.com/docs/books/jvms/second_edition/ClassFileFormat-Java5.pdf

A fix for this was applied to ASM, and this upgrade pulls it in. Classfile verification during compile is now re-enabled.