Fix JRUBY-5417: Nil backtrace when overriding method_missing and calling super
Ok, the problem here is that with the "backtrace" branch work, it became a requirement that RaiseException must have its preRaise method called before getting thrown, so it would generate the Java backtrace. This matched Ruby semantics for Exception.new/raise lifecycle, but put the onus on Java-land exception creators to just know they had to make that call. In practice, nobody knows that, and even in JRuby several extensions were still not preRaise'ing.

This particular case was caused by a separate logical path for method_missing fired in the "super" case. That path did not preRaise.

I've decided to just force the preRaise call into the RaiseException constructors, so it is guaranteed to be there. This passes all tests and specs, and will fix this error plus many others as yet unreported. It also still mostly matches the Ruby lifecycle, since Exception.new does not create a RaiseException in Java...that doesn't happen until raise is actually called anyway.