Fix JRUBY-5412: Strange each_slice/with_index combo exception
The problem here was that we were setting the block passed to the each_slice enumerator's each as having arity two, when it actually only had arity one (because it's each_with_index, which internally is arity one and becomes arity two by adding the additional argument. This triggered an optimized path in each_slice that re-used the same array for each iteration, ultimately producing an unusable array at the end and causing the explosion.

I have repaired the bad arity, but I may also revert the optimization and revisit it in the future, since it's obvious that bad arities (of which we have many in JRuby) could trigger future failures.