Fix JRUBY-5418: Getting NativeException: org.jruby.RubyContinuation$Continuation: null with RSpec 2 and EventMachine
Java integration logic catches Java exceptions raised in the call and rethrows them as NativeException. It also specially catches RaiseException so it can propagate through as-is, but it also needed to do the same for "Unrescuable" exceptions like those used for continuations (catch/throw/callcc), flow control, system exit, thread death, and so on. I modified the logic to additionally re-throw unmodified any "Unrescuable" exceptions.