Fix JRUBY-5483: become_java! only works on the first subclass of a deep class hierarchy Fix JRUBY-5491: jruby.reify.classes fails with optparse.rb (trinidad)
The logic for when a class is reifiable was incorrect; it only allowed the first unreified chain of classes up to Object to reify; other children of reified classes would fail to reify and descendants of core classes other than Object or BasicObject were allowed to reify. The new logic allows only classes meeting the followign criteria to reify:

* not already reified
* reified parent must be Object, BasicObject, or Reified impl (new marker interface)
* unreified parent must be reifiable

This issue also caused JRUBY-5491 because classes that should be reified were not, resulting in them keeping the reifiable object allocator and recursing forever.