Fix JRUBY-5471: private method `verify_mode=' called with Bundler, net/https, jruby-openssl and RubyGems 1.5
The problem here was that the extension loaded with the requirer's frame stil present on the stack. If the requirer had set visibility to "private", frame-sensitive methods like attr_accessor would use that for the method visibility. In MRI, extensions are most likely given a fresh frame in which to execute. We did not see this before because most core JRuby methods used to be framed (like Kernel.require) and so extensions implicitly got a clean environment.

The update to RubyGems probably exposed this bug (likely present in RC1 as well) by setting OpenSSL attributes where it had not done so before.

The fix is to push a fresh frame before loading extensions, so they execute in a clean environment with public visibility default.