Fix JRUBY-4469: Process.spawn seems to be completely broken
Process.spawn appeared to be implemented like popen, which meant it was expecting streams to be driven by user code. I modified it to launch the process and "pump" the out and err streams, leaving the in stream alone.

Note that this is not quite like MRI yet, since through normal Java process-launching mechanisms we can't launch a process that inherits stdio streams from the parent (at least, not until Java 7). So the streams are being "pumped", one thread each for out and err, resulting in the output being a bit stuttery.

The best solution would be for the current behavior to be our fallback "pure Java" logic, with the appropriate native calls used when possible.