Fix JRUBY-4788: Stackoverflow when calling code from java with wrong parameters
When dispatching to an abstract method, we checked if the Ruby arity matched the incoming arguments. If it did not, we unconditionally redispatched against the superclass. If the superclass method was abstract, this would pick up the JavaProxyMethod stub, dispatching again back through the mismatched-arity logic. This caused a stack overflow. By checking whether the superclass has a dispatchable method, we can properly present an ArgumentError instead.