Fix JRUBY-5638: jrubyc path output
After investigating a bit I could find no indication that the mangled class packages were even loadable by JRuby in any normal fashion. Logically, what people usually want from a command-line compilation step is to precompile all .rb files to .class files in the same path structure. So I have modified the compiler to do just that. The in-.class name will use a SHA1 hash of the original source to guarantee uniqueness. The .class files themselves and the paths into which they are emitted will be named as the originals.

The additional scenario (which has never really been utilized) is precompiling to allow preverification, conversion to Dalvik bytecode, and so on. This mode will likely require a separate compiler option to emit not just .class files but a fully-compiled codebase that manages paths, invokers, and classloading appropriately. jrubyc currently just does not support this case, so we narrow its features to emitting precompiled ruby code for typical distribution and generating Java classes and source files when creating Java types from Ruby scripts.