Update StringValuePtr/rb_string_value impl to match MRI.
This fixes an issue in the nokogiri extension when passing in a
BlankSlate object that pretends to be a string. #kind_of? and #class
claim the object is a string, even though it is not an RString.

Breakpoint 2, jruby_str (v=4296932608) at /code/jruby/cext/src/string.cpp:465
465         rb_raise(rb_eTypeError, "wrong type (expected String)");
(gdb) where
 #0  jruby_str (v=4296932608) at /code/jruby/cext/src/string.cpp:465
 #1  0x000000010c7834b3 in jruby_str_cstr (v=4296932608) at /code/jruby/cext/src/string.cpp:474
 #2  0x000000010c783636 in rb_string_value_ptr (object_variable=0x1005fc660) at /code/jruby/cext/src/string.cpp:397
 #3  0x000000010c109f4b in read_memory (klass=4296361264, string=4296932608, url=4, encoding=4, options=4291) at html_document.c:91

Signed-off-by: Tim Felgentreff <timfelgentreff@gmail.com>