Fix JRUBY-5871: java.lang.NegativeArraySizeException from RubyEnumerator (after JITed)
The problem here was that the compiler did not properly route "special args" (like foo *args) to the specific-arity path they should go to. This affected any such calls as well as super and zsuper calls, though in this case it was not the latter two cases. The fix is to add a new call path for "varargs" that checks arg arity and reroutes along the right path. I also fixed Enumerator#initialize's varargs path to ensure at least 3 args, which would have prevented a Java error in the first place (but replaced it with an ArgumentError.