Support byte limits in IO#gets and StringIO#gets in 1.9 mode
In 1.9 mode, IO#gets should be able to take a maximum length either as the only argument, or as the second argument (with the first being nil or a separator). The current implementation throws an ArgumentError, TypeError, or NPE when you pass in a limit, depending on how you call it.

This patch adds basic support to IO and StringIO for the limit argument, but some more work is probably needed for multibyte separators (e.g., deciding what to do with the line count when the separator straddles the user's passed-in byte limit).

Signed-off-by: Charles Oliver Nutter <headius@headius.com>