Allow disabling ENV write-through to the real environment [JRUBY-5934]
This adds updateNativeENVEnabled to RubyInstanceConfig with a default of true.
If it is true and jruby.native.enabled is also true, any modifications to
the runtime's ruby ENV will be reflected in the system env via POSIX. Otherwise,
the modifications are not written through.

This setting has not yet been pushed out to the embedding API's (ScriptingContainer
& JRubyEngine), nor has it been documented.