Move RString and RArray native handles to varTable.
RString and RArray are cached in RubyString and RubyArray to avoid weak-caching
handles or recreating handles frequently. However these fields add an additional
32 or 64 bits to the size of all String and Array instances even when C exts are
not in use. Since C ext use on JRuby is very uncommon, I've moved these fields
into the lazily-allocated variable table, in the same manner as the rarely-used
object_id. This reduces the size and allocation cost of all String and Array
instances for the good of all JRubyists.

The cost of retrieving the handles increases slightly due to hopping through
getMetaClass.getRealClass, the accessor, and the array, but that cost likely
pales in comparison to the already-high cost of C extensions, and only impacts
Array and String instances that require native handles.