Enable switchpoint call site invalidation and indirect call paths for invokedynamic support.
SwitchPoint-based class hierarchy invalidation previously invalidated each in turn. This was
known to be inefficient, since each invalidation forces the JVM to a safepoint in order to
push invalidations out to code that has inlined through a SwitchPoint. I modified the
invalidation logic to aggregate a collection of all SwitchPoints from child classes/modules
and invalidate them as a batch.

My benchmark was the time it takes to run the FFI specs. Because they run in rspec, and
rspec redefines top-level methods for each example, ths SwitchPoint invalidation is hit hard.
With the modified logic, the SP-based invalidation slows the suite to about 2x what it is
without SP invalidation. This can still be improved, but it's good enough to enable SP
invalidation for now.