Fix for JRUBY-6101.
ScriptingContainer as well as JSR223 JRubyEngine uses non-autoclose RubyIO to redirect standard i/o. I tried to figure out the good place to close those non-autoclose IO with no luck. This is because of the presence of finalize() method. This might cause some trouble, but for now, let them be GCed and no close explicitly.