Prevent cache leakage in extreme method churn situations.
While working on a support ticket, it came to my attention that Rails was
generating lots of new methods for templates due to some gap in caching.
These methods were generated with a new name each time, using the __id__
of the Template object that would eventually manage them.

Because our per-class caching did not actively flush (only passively,
by invalidating each class and letting subsequent calls recache), old
method names never encountered again would stick around in the cache.
This led to a leak, as more and more new method names were created,
called, and then abandoned.

By actively invaldating the per-class cache in response to an invalidation
event, we avoid the leak. It does not appear to add significant overhead
to the invalidation process.