Tag obscure autoload failures.
These failures were brought to light by fixing how we calculate the
new name for anonymous classes assigned to a new constant in a
module/class container. Because we did not recalculate before, the
autoload was found under the old anonymous name. Now that we
recalculate (correctly) the autoload can't be located.

If this gets reported as a real user-visible problem, we can look
into a fix, but it's fairly obscure and 1.7 does autoloads in a new
way anyway.