Remove class name from backtrace data.
Class name was only used for Rubinius backtraces, which nobody is
using. JRuby's default backtrace is now very similar, so I've
removed the Rubinius format completely. Also, for JRUBY-6218, we
regressed on interpreter performance because class names are no
longer cached. Also, in compiled Ruby code, the class names are not
available most of the time. All this combined to convince me to just
remove class name from the backtrace data.

The "rubinius" backtrace style setting still works, but does not
change anything (just uses the "jruby" style).

This fixes the perf regression for the interpreter, but does not
do anything to generally fix the regression for Class/Module name
calculation performance.