Add a config property to enable a single global require lock.
The ability to require files in parallel in Ruby leads to many
threading bugs. Constants can be defined before they're ready for
use, classes can be modified while they're being used, global
state can be initialized at the same time it's in use. Bad news.

This flag is experimental, and forces all requires to happen in
in serial. First thread to acquire the lock keeps it until it
returns from require. This prevents all potential threading bugs.

It is not enabled by default, in case there's libraries out there
that require and never return. For example, if require has a side
effect of starting an event loop, and so it is usually required
within a new thread. That pattern will have to go away in order
for this flag to be useful (and by extension, for requires to be
made properly threadsafe).