Reduce case/when fast-path optimization.
The old optimization did not consider hash collisions on strings and
symbols, and would not work if we randomize string hashcodes. Rather
than try to hack around those issues, I'm just reducing the string
and symbol fast paths to the single-character form only. The ===
call is not free, but we can revisit the hashcode-based
optimization at a later date and do it right then.