A new implementation of ThreadFiber that passes specs.
The raw LockSupport use I had before was fast, but I did not deal
with the fact that any fiber could end up waking any thread,
regardless of whether that thread was waiting on the fiber. By
switching to Exchanger, a given fiber can only wake the thread
that is actually waiting on it.

This may not be as fast as the other impl, but works properly. I
will reinvestigate performance.