[SECURITY] Random String#hash to avoid DoS attack
This is the fix for CVE-2011-4838.

Randomize String#hash (in Ruby world) and internal RubyString#hashCode()
(in Java world) for 'Denial of Service via Algorithmic Complexity
Attacks' explained in [1].  Introduce the runtime-local seed for hash
randomization which is set at runtime initialization and use it as the
seed for MurmurHash 2.0 [2].  MurmurHash is used in CRuby 1.9 and it
should make the attacks difficult. (Assuming that no reverse function
has been found for MurmurHash 2.0)

Credit to Alexander Klink <alexander.klink@nruns.com> and
Julian Waelde <jwaelde@cdc.informatik.tu-darmstadt.de> for reporting
this issue.

[1] http://www.cs.rice.edu/~scrosby/hash/CrosbyWallach_UsenixSec2003.pdf
[2] http://sites.google.com/site/murmurhash/