Remove unnecessary locking on var table write access.
The varTable we use to store internal and instance variables is a
volatile field. Previously, in order to guarantee two threads growing
that field did not step on each other, we synchronized around the
grow logic. This unfortunately also meant that cases which required
no growth would also be forced to acquire a lock.

By switching the update to a CAS loop using AtomicFieldUpdater, we
avoid the lock and only spin when there's contention. If no growth
is required, the array is returned immediately. If growth or init
is required, we do so atomically, to make a best effort at avoiding
stepping on another thread's toes.

Note that this change does nothing to guarantee the atomicity of
variable writes or to guarantee that multiple threads will not miss
each others' writes into the variable table, but the race conditions
only apply early in execution during the very first writes to
instance vars, when the class has not yet settled into a steady-state
table size. Once there are no new instance vars being assigned, the
table is allocated once, atomically, using the steady-state size.

Note also that individual instance variables are still not atomic
or volatile. If you need an atomic reference, use the 'atomic' gem.