Set English.rb aliases immediately, and make it a no-op.
Because we rely on the presence of $~ and $_ global variables to
know we must allocate space for them, we need to define their
English.rb aliases right away and disallow further aliasing. Since
we must define those aliases, we might as well define all aliases
and make English.rb a no-op.