Hack for popen on OpenJDK 7.
OpenJDK 7 appears to kill off the file descriptor for the child process
when it exits, but oddly enough the channel does not get marked closed.
This leads to an EBADF style error when we go ahead and try to close
what looks like an open channel. By checking for that error and
confirming the process is dead, we can ignore the spurious error.