Remove ruby_test suite.
The ruby_test suite originally started as a suite of tests for
standard Ruby functionality in the "facets" library of core
extensions. We imported a copy of it many years ago to increase our
test coverage, but since then our copy and the real copy have steadily
diverged. The actual suite is now more focused on testing the core
extensions in the library, while ours has remained a standard Ruby
functionality test with modifications where we needed to mask out
impl-specific behavior or bad tests.

At this point, without 1.9 support in ruby_test and with RubySpec
and MRI's own tests providing better feature coverage, I've opted
to remove the ruby_test suite.