Fix JRUBY-6281: [1.9] Applet does not work in the 1.9 mode
For 1.9 mode we canonicalize loaded paths to make sure we are not
double-loading files that are required in different ways, and so we
can properly use per-file locking around requires. However, in a
restricted environment we may not be able to canonicalize. My fix
falls back on just using the literal require name in such cases.

I also split the test-security-manager task into 1.8 and 1.9 versions.

Conflicts:

	build.xml