Fix JRUBY-6280: Fails to open fifo for writing.
It turns out that in glibc, fopen operations that would seek in
the resulting file descriptor do actually check for ESPIPE to know
that a failed seek is due to the file being a pipe or FIFO. We do
not have direct access to the actual ESPIPE, but we can check the
IOException message to know if it's the same problem.

I don't think there's any real risk in this change, since it will
only quietly fail to seek on a file when it's impossible to seek
on that file, which is what we want. Other error cases and open
failures should propagate as before.

Conflicts:

	default.build.properties
	src/org/jruby/util/io/ChannelDescriptor.java