[IR] Fixed a long-standing zsuper bug by introducing a ZSuperInstr that can handle compile-time unknown zsuper-args scenario.
YET TO BE DONE: Before a zsuper instruction, add dummy instructions that
use all block/method args of all scopes in the scope nesting tree
starting with the scope in which zsuper occurs till we hit a method
scope.  This is needed to ensure that method/block args are not eliminated
as part of dead-code removal.  Without this, we might end up deleting
args which a zsuper might subsequently need.