Fix JRUBY-5630: Java calls have lost the Java part of the trace
I fixed this by having normal Ruby traces flip to an "integrated"
mode when handling JI-sourced exceptions. Integrated mode will
include all Java frames that aren't org.jruby, sun.reflect, or
otherwise handled by interpreter/compiler demangling. This allows
the target Java call's trace to show up at the top of the Ruby
trace.

This is not quite perfect; I'll need to revisit this in the future
to make sure we're not allowing too much in. For example, several
of our shipped libraries will not be filtered, so if they're in
the trace they'll show up interspersed with Ruby frames. This may
be rare, however, since almost none of those libraries are call-
through cases...they're usually leaf code.