Fix JRUBY-6325: Java::OrgYamlSnakeyamlReader::ReaderException: special characters are not allowed
I do not have a clear explanation for why this was failing before,
but by using an InputStreamReader with the charset from the Ruby
string, rather than calling toString, we are able to deal with
characters from the input strema. I will investigate further why
the toString path did not work properly, but this fixes the issue.

Can someone stick a test case into JRuby's regression tests or
into RubySpec?