Re-fix regression spec about cross runtime Object handling
You cannot do id1.should == id2 here because the Objects are from
different runtime!  Cross runtime comparison confuses rspec and it tries
to raise an expectation exception, but the RaiseException is also
confusing about runtime so it can be thrown.

Cross runtime object passing is evil.  I should implement Channel for
this purpose.