Fix JRUBY-4913
FileUtils.cp failed if file source is too big (>= 1GB)

On 32-bit JVMs, FileChannel.transferTo would attempt to mmap a
a larger addressable space than the process could handle. This
caused the error. My fix is to catch that error and then attempt
the copy again in 100MB chunks. 100MB may still fail if we are
close to the process's addressable memory limit, but it's much
more reliable than what we had before.