Good-enough fix for JRUBY-6341
Regression: Enumerable#first on a Java list returns an array, not an element

My fix is to look for single-arg calls and use that arg rather than
boxing everything in a new array. It may not be correct but it fixes
this case and breaks nothing else We'll revisit this before 1.7...
we kinda have to.