Fix Scala singleton support.
Scala singletons were being walked for their companion methods, but
because we only add methods that are "local" under normal
circumstances (local meaning on that class or on an interface) we
never actually installed them. Modified the installer to always
install Scala singleton methods and added/reworked specs to do the
right thing.