Fix GzipReader encoding on 1.9 mode (produced encoding == null String)
It's a fix for "-J-ea pointed out we were creating a bad ByteList with
no encoding in 1.9 mode..." reported by Tom.

The actual bug is from 1 line, that I forgot to set default
externalEncoding to runtime.getDefaultExternalEncoding()
But the problem is from name confusion.  Use proper names, readEncoding
and writeEncoding for enc, enc2 in Ruby's IO subsystem.