Always grab switchpoint before looking up methods/constants.
Because switchpoint invalidation always happens *after* constant
and method tables have been updated, we must work the opposite
direction in the call sites to ensure if a modification happens,
we will be an invalid switchpoint in hand. This is akin to the JVM
happens-before guarantee:

* Method/constant table update happens-before switchpoint invalidation
* SwitchPoint retrieval happens-before method/constant lookup

Reversing the order at call sites can cause a data race, where the
value (constant/method) in hand may get cached with a subsequent
update's switchpoint, leading to a stuck cache until the next
update.