Fix JRUBY-6522
ThreadLocal RubyEncoding.UTF8_CODER leaks class loader

Because our threadlocal for UTF-8 encoding/decoding references a
JRuby-specific class, it leaks classloaders when redeploying in a
server environment that isolates redeploys in their own
classloaders. Indead of a hard thread-local reference, this fix
uses a soft reference, so the contents will age out or be cleaned
immediately when there's memory pressure.