Fix JRUBY-6528
Socket#connect_nonblock and IO::select appear to be misbehaving?

When selecting for "write" in Ruby terms, we need to select for
both "write" and "connect" in Java selector terms. On most impls
of selectors, there's only "read" and "write" to be used for
read/accept and write/connect. We were only registering interest
in the Java "write" event", but the JDK knows we're doing select
against an unconnected socket, so it's looking for "connect"
interest. It wakes up, because at the native level "write" and
"connect" are both just "write", sees that we're not interested
in "connect", and we don't end up adding anything to the outgoing
writable array.

By modifying "write" selection events to do write + connect (and
fixing "read" selection events to do read + accept) the given
script works as expected.

This code still needs some improvement and refactoring, but I'm
closer to understanding how to do it right now.