Fix JRUBY-6570: autoload :Time, 'time' doesn't work
Instead of always using Object when calling autoload from an
instance method, we use self's real class (skipping singletons)
as the autoload location. This isn't exactly how MRI does it
(they use cbase, which I think is euivalent to 1.8's ruby_class),
but it should be correct for the majority of cases.